requirements.txt
flask==2.3.3
google-generativeai==0.3.1
python-dotenv==1.0.0
pandas==2.0.3
PyPDF2==3.0.1
numpy==1.24.3
gunicorn==21.2.0
flask-cors==4.0.0
werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
blinker==1.6.2
python-pptx==1.0.2
pillow==11.1.0
PyMuPDF==1.23.7
chardet==5.2.0
httpx==0.26.0
asyncio==3.4.3
